Nuevo commit

+++ Controllers/CitasController.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using Proyecto_Programacion_III.Data;
 using Proyecto_Programacion_III.Models.Entidades;
+using Proyecto_Programacion_III.Models.Entidades.Opciones;
 
 public class CitasController : Controller
 {
@@ -17,7 +18,8 @@ public class CitasController : Controller
     {
         var citas = _context.Citas
             .Include(c => c.Cliente)
-            .Include(c => c.Servicio);
+            .Include(c => c.Servicio)
+            .Include(c => c.Usuario); 
 
         return View(await citas.ToListAsync());
     }
@@ -27,6 +29,7 @@ public class CitasController : Controller
     {
         ViewBag.Clientes = _context.Clientes.ToList();
         ViewBag.Servicios = _context.Servicios.ToList();
+        ViewBag.Usuarios = _context.Usuarios.ToList(); 
         return View();
     }
 
@@ -34,25 +37,100 @@ public class CitasController : Controller
     [HttpPost]
     public async Task<IActionResult> Create(Cita cita)
     {
+        cita.Estado = EstadoCita.Programada;
+        if (cita.FechaHora < DateTime.Now)
+        {
+            ModelState.AddModelError("FechaHora", "No se pueden agendar citas en fechas pasadas");
+        }
+        var servicio = _context.Servicios
+        .FirstOrDefault(s => s.ServicioId == cita.ServicioId);
+
+        if (servicio != null && servicio.Estado == EstadoServicio.Inactivo)
+        {
+            ModelState.AddModelError("ServicioId", "El servicio se encuentra inactivo");
+        }
+
         if (ModelState.IsValid)
         {
-            _context.Add(cita);
+            _context.Citas.Add(cita);
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
         }
 
         ViewBag.Clientes = _context.Clientes.ToList();
         ViewBag.Servicios = _context.Servicios.ToList();
+        ViewBag.Usuarios = _context.Usuarios.ToList();
+
+        return View(cita);
+    }
+    public async Task<IActionResult> Edit(int id)
+    {
+        var cita = await _context.Citas.FindAsync(id);
+
+        if (cita == null)
+            return NotFound();
+
+        if (cita.Estado == EstadoCita.Cancelada)
+            return RedirectToAction("Index");
+
+        ViewBag.Clientes = _context.Clientes.ToList();
+        ViewBag.Servicios = _context.Servicios.ToList();
+        ViewBag.Usuarios = _context.Usuarios.ToList();
+
         return View(cita);
     }
 
+    [HttpPost]
+    public async Task<IActionResult> Edit(Cita cita)
+    {
+        var citaDb = await _context.Citas.FindAsync(cita.Id);
+
+        if (citaDb == null)
+            return NotFound();
+
+        if (citaDb.Estado == EstadoCita.Cancelada)
+        {
+            ModelState.AddModelError("", "No se puede editar una cita cancelada");
+        }
+
+        if (cita.FechaHora < DateTime.Now)
+        {
+            ModelState.AddModelError("FechaHora", "No se pueden usar fechas pasadas");
+        }
+
+        var servicio = _context.Servicios
+            .FirstOrDefault(s => s.ServicioId == cita.ServicioId);
+
+        if (servicio != null && servicio.Estado == EstadoServicio.Inactivo)
+        {
+            ModelState.AddModelError("ServicioId", "Servicio inactivo");
+        }
+
+        if (ModelState.IsValid)
+        {
+            citaDb.ClienteId = cita.ClienteId;
+            citaDb.ServicioId = cita.ServicioId;
+            citaDb.UsuarioId = cita.UsuarioId;
+            citaDb.FechaHora = cita.FechaHora;
+
+            await _context.SaveChangesAsync();
+            return RedirectToAction("Index");
+        }
+
+        ViewBag.Clientes = _context.Clientes.ToList();
+        ViewBag.Servicios = _context.Servicios.ToList();
+        ViewBag.Usuarios = _context.Usuarios.ToList();
+
+        return View(cita);
+    }
 
     public async Task<IActionResult> Delete(int id)
     {
         var cita = await _context.Citas
             .Include(c => c.Cliente)
             .Include(c => c.Servicio)
-            .FirstOrDefaultAsync(c => c.CitaId == id);
+            .Include(c => c.Usuario)
+            .FirstOrDefaultAsync(c => c.Id == id); 
 
         return View(cita);
     }
@@ -61,8 +139,13 @@ public class CitasController : Controller
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
         var cita = await _context.Citas.FindAsync(id);
-        _context.Citas.Remove(cita);
-        await _context.SaveChangesAsync();
+
+        if (cita != null)
+        {
+            _context.Citas.Remove(cita);
+            await _context.SaveChangesAsync();
+        }
+
         return RedirectToAction("Index");
     }
 }
